style: 유저 입력값에 대한 유효성 검사 코드 분리
관리를 위해 유효성 검사 관련 코드 Validate객체로 이동
